changed latitude and longitude to now be floats

+++ trails/db/migrate/20151126003733_create_trails.rb
@@ -3,10 +3,10 @@ class CreateTrails < ActiveRecord::Migration
     create_table :trails do |t|
       t.string :Name
       t.string :Description
-      t.integer :StartLng
-      t.integer :StartLat
-      t.integer :EndLng
-      t.integer :EndLat
+      t.float :StartLng
+      t.float :StartLat
+      t.float :EndLng
+      t.float :EndLat
 
       t.timestamps null: false
     end
